Continue working on drivetrain code to drive with x-box controller
