+++ src/main/java/frc/robot/commands/DriveRobotWithJoysticks.java
@@ -1,0 +1,48 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
+
+package frc.robot.commands;
+
+import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.Robot;
+import frc.robot.subsystems.DriveTrainSubsystem;
+
+public class DriveRobotWithJoysticks extends CommandBase {
+  /**
+   * Creates a new DriveRobotWithJoysticks.
+   */
+    // The subsytem command runs on the following:
+    private final DriveTrainSubsystem m_drivetrain_subsystem;
+
+  public DriveRobotWithJoysticks(DriveTrainSubsystem subsystem) {
+    // Use addRequirements() here to declare subsystem dependencies.
+    m_drivetrain_subsystem = subsystem;
+    addRequirements(m_drivetrain_subsystem);
+  }
+
+  // Called when the command is initially scheduled.
+  @Override
+  public void initialize() {
+  }
+
+  // Called every time the scheduler runs while the command is scheduled.
+  @Override
+  public void execute() {
+      m_drivetrain_subsystem.move_with_joysticks(Robot);
+  }
+
+  // Called once the command ends or is interrupted.
+  @Override
+  public void end(boolean interrupted) {
+  }
+
+  // Returns true when the command should end.
+  @Override
+  public boolean isFinished() {
+    return false;
+  }
+}
