+++ src/main/java/frc/robot/subsystems/DriveTrainSubsystem.java
@@ -6,6 +6,7 @@ package frc.robot.subsystems;
 
 import com.ctre.phoenix.motorcontrol.*;
 import com.ctre.phoenix.motorcontrol.can.*;
+import edu.wpi.first.wpilibj.XboxController;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
@@ -21,6 +22,10 @@ public class DriveTrainSubsystem extends SubsystemBase {
   private final WPI_VictorSPX leftMaster = new WPI_VictorSPX(4); // This is the CAN ID for the device
   private final WPI_VictorSPX leftSlave = new WPI_VictorSPX(2); // This is the CAN ID for the device
 
+  // Definition of joystick axes variables for left joystick
+  public final int left_x_axis = 0;
+  public final int left_y_axis = 1;
+
   private final DifferentialDrive drive = new DifferentialDrive(leftMaster, rightMaster);
 
   public DriveTrainSubsystem() {
@@ -28,8 +33,13 @@ public class DriveTrainSubsystem extends SubsystemBase {
     leftSlave.follow(leftMaster);
   }
 
-  public void move(double xSpeed, double zRotation) {
-    drive.arcadeDrive(xSpeed, zRotation);
+  public void move_with_joysticks(XboxController driver_controller) {
+
+    // Get axis values for speed and rotational speed
+    double xSpeed = driver_controller.getRawAxis(left_x_axis);
+    double zRotation_rate = driver_controller.getRawAxis(left_y_axis);
+
+    drive.arcadeDrive(xSpeed, zRotation_rate);
   }
 
   @Override
